# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arena Protocol API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LeaderboardEntry {
  rank: number;
  address: string;
  totalWins: number;
  totalBattles: number;
  totalRewards: string;
  winRate: number;
}

export interface PlayerStats {
  address: string;
  totalWins: number;
  totalBattles: number;
  totalRewards: string;
  fighters: number;
  rank: number;
}

export interface BattleRecord {
  id: number;
  player: string;
  fighterId: string;
  win: boolean;
  reward: string;
  mode: string;
  txHash?: string;
  timestamp: string;
}

export interface RecordBattleRequest {
  player: string;
  fighterId: string;
  win: boolean;
  reward: string;
  mode: string;
  txHash?: string;
}

export interface MarketListing {
  tokenId: string;
  seller: string;
  price: string;
  rarity: string;
  strength: number;
  speed: number;
  intelligence: number;
  wins: number;
  losses: number;
  active: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type GetLeaderboardParams = {
  limit?: number;
};

export type GetBattleHistoryParams = {
  limit?: number;
  address?: string;
};

export type GetMarketListingsParams = {
  limit?: number;
};