# lib/api-zod/src/generated/types/marketListing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arena Protocol API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MarketListing {
  tokenId: string;
  seller: string;
  price: string;
  rarity: string;
  strength: number;
  speed: number;
  intelligence: number;
  wins: number;
  losses: number;
  active: boolean;
}