# lib/api-zod/src/generated/types/playerStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arena Protocol API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerStats {
  address: string;
  totalWins: number;
  totalBattles: number;
  totalRewards: string;
  fighters: number;
  rank: number;
}