# lib/api-zod/src/generated/types/recordBattleRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arena Protocol API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecordBattleRequest {
  player: string;
  fighterId: string;
  win: boolean;
  reward: string;
  mode: string;
  txHash?: string;
}